sink(ticdc): remove mysql sinkv1
